Fix database access caused by calling primary_key
This fixes a regression caused by the change to allow customization of
the primary key. While calling `primary_key` on the model is the proper
way to find it's pk, it caused a schema lookup which should be avoided
in class methods so things like asset precompilation can be run without
a database.

This fixes #54.